requirements.txt
amara3.iri==3.0.3
amara3.xml==3.2.1
anyio==3.3.1
appnope==0.1.2
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
bookops-worldcat==0.4.0
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
clean-text==0.5.0
click==6.7
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
emoji==1.5.0
entrypoints==0.3
et-xmlfile==1.1.0
frozendict==2.0.7
ftfy==6.0.3
html5lib==1.1
idna==3.2
iniconfig==1.1.1
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
isodate==0.6.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==3.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-core==4.8.1
jupyter-server==1.11.0
jupyterlab==3.1.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.1
lxml==4.6.3
marc2bib==0.1.4
marctools==0.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mmh3==3.0.0
nameparser==1.0.6
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.4
openpyxl==2.6.4
packaging==21.0
pandoc==1.1.0
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pluggy==1.0.0
plumbum==1.7.0
ply==3.11
prometheus-client==0.11.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pybibframe==1.1.4
pycparser==2.20
Pygments==2.10.0
PyICU==2.7.4
PyLD==2.0.3
pymarc @ file:///Users/andrewc/dev/i18n/libr/pymarc
pyparsing==2.4.7
pyrsistent==0.18.0
PySide2==5.15.2
PySimpleGUI==4.55.1
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.1
PyZ3950 @ file:///Users/andrewc/dev/i18n/libr/adlam-test/z/asl2_PyZ3950
pyzmq==22.3.0
rdflib==6.0.2
regex==2021.9.24
remi==2021.3.2
requests==2.26.0
requests-unixsocket==0.2.0
Send2Trash==1.8.0
shiboken2==5.15.2
six==1.16.0
sniffio==1.2.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
unicodedataplus==13.0.0.post2
Unidecode==1.3.2
urllib3==1.26.6
versa==0.5.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
wxPython==4.1.1
